再次重构，加入相机移动和缩放的逻辑

+++ Assets/Scripts/Data/AstroData.cs
@@ -51,7 +51,7 @@ public struct AstroData : IPoolElement
         this = default;
     }
 
-    public void InternelUpdate(float deltaTime)
+    public void InternelUpdate(double deltaTime)
     {
         if (!active)
             return;
